Moved all utilities to new location
