+++ IPA.Loader/Utilities/LoneFunctions.cs
@@ -5,7 +5,7 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace IllusionInjector.Utilities
+namespace IPA.Utilities
 {
     public static class LoneFunctions
     {
